// packages/twenty-server/src/modules/executive-search/common/enums/search-assignment-status.enum.ts
export enum SearchAssignmentStatus {
  BD_HANDOFF = 'BD_HANDOFF',
  CONTRACTING = 'CONTRACTING',
  KICKOFF = 'KICKOFF',
  ACTIVE = 'ACTIVE',
  ON_HOLD = 'ON_HOLD',
  PLACEMENT_PENDING = 'PLACEMENT_PENDING',
  PLACED = 'PLACED',
  GUARANTEE = 'GUARANTEE',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  LOST = 'LOST',
}